refactor(error): centralize anyhow envelope parts mapping
