+++ openspec/changes/refactor-anyhow-error-parts-helper/.openspec.yaml
@@ -1,0 +1,2 @@
+schema: spec-driven
+created: 2026-01-23
