+++ openspec/changes/refactor-anyhow-error-parts-helper/README.md
@@ -1,0 +1,3 @@
+# refactor-anyhow-error-parts-helper
+
+Centralize `anyhow::Error` -> `(code, message, details)` mapping for error envelopes.
